+++ examples/python/ssa_tests/ssa_testi.py
@@ -1,6 +1,6 @@
 #! /usr/bin/env python
 #
-# Copyright (C) 2011, 2012, 2013, 2014, 2015, 2016 Ed Bueler and Constantine Khroulev and David Maxwell
+# Copyright (C) 2011, 2012, 2013, 2014, 2015, 2016, 2018 Ed Bueler and Constantine Khroulev and David Maxwell
 #
 # This file is part of PISM.
 #
@@ -106,12 +106,13 @@ class testi(PISM.ssa.SSAExactTestCase):
 # The main code for a run follows:
 if __name__ == '__main__':
     context = PISM.Context()
+    config = context.config
 
     PISM.set_abort_on_sigint(True)
 
-    Mx = PISM.optionsInt("-Mx", "Number of grid points in x-direction", default=11)
-    My = PISM.optionsInt("-My", "Number of grid points in y-direction", default=61)
-    output_file = PISM.optionsString("-o", "output file", default="testi.nc")
+    config.set_double("grid.Mx", 11, PISM.CONFIG_DEFAULT)
+    config.set_double("grid.My", 61, PISM.CONFIG_DEFAULT)
 
-    tc = testi(Mx, My)
-    tc.run(output_file)
+    tc = testi(int(config.get_double("grid.Mx")),
+               int(config.get_double("grid.My")))
+    tc.run(config.get_string("output.file_name"))